+++ src/main/java/racingcar/controller/CarRacingWebController.java
@@ -8,9 +8,9 @@ import org.springframework.validation.FieldError;
 import org.springframework.validation.ObjectError;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.*;
-import racingcar.dto.RacingResultDTO;
-import racingcar.dto.RacingStartDTO;
 import racingcar.service.CarRacingService;
+import racingcar.service.dto.RacingResultDTO;
+import racingcar.service.dto.RacingStartDTO;
 
 import javax.validation.Valid;
 import java.util.HashMap;
