+++ src/test/java/racingcar/controller/CarRacingWebControllerTest.java
@@ -15,7 +15,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
-import racingcar.dto.RacingStartDTO;
+import racingcar.service.dto.RacingStartDTO;
 
 import java.util.List;
