+++ src/test/java/racingcar/dao/JdbcGameDaoTest.java
@@ -8,7 +8,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.jdbc.core.JdbcTemplate;
-import racingcar.dto.GameIdDTO;
+import racingcar.dao.dto.GameIdDTO;
 
 import java.util.List;
