+++ src/test/java/racingcar/dao/LocalGameDaoTest.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
-import racingcar.dto.GameIdDTO;
+import racingcar.dao.dto.GameIdDTO;
 
 import java.util.List;
